requirements.txt
# Core dependencies
pinecone-client>=3.0.0
oracledb>=2.0.0
python-dotenv>=1.0.0

# LLM and Embeddings
openai>=1.12.0
anthropic>=0.18.0
langchain>=0.1.0
langchain-openai>=0.0.5
tiktoken>=0.5.2

# Text processing
beautifulsoup4>=4.12.0
python-docx>=1.1.0
pypdf>=4.0.0
markdown>=3.5.0

# Data processing
pandas>=2.0.0
numpy>=1.24.0

# Utilities
pydantic>=2.5.0
pydantic-settings>=2.1.0
loguru>=0.7.0
tenacity>=8.2.0
tqdm>=4.66.0
